Fixed an issue with the components macro paths
The paths used inside the `generate_components_simmechanics` macro were pointing to the files, but the paths were not set with CMake and it therefore could not find the files. This has been fixed by adding the `set` and `configure_files` commands before the `add_custom_command`.

Also, we included the copying of the meshes and configuration files for the left_hand_mk3. If/when the wrist is included, the copy commands for the wrist should be included as well.